Move JsonApiAssertions helper class into project source
